+++ .eslintrc.js
@@ -467,13 +467,13 @@ module.exports = {
 			extends: [ 'plugin:ssr-friendly/recommended' ],
 		},
 		{
-			files: [ 'packages/components/src/**' ],
+			files: [ 'packages/components/src/**', 'packages/ui/src/**' ],
 			rules: {
 				'no-restricted-imports': [
 					'error',
 					// The `ariakit` and `framer-motion` APIs are meant to be consumed via
-					// the `@wordpress/components` package, hence why importing those
-					// dependencies should be allowed in the components package.
+					// the `@wordpress/components` and @wordpress/ui` packages, hence why
+					// importing those imports should be allowed only in those packages.
 					{
 						paths: restrictedImports.filter(
 							( { name } ) =>