+++ packages/block-editor/src/components/url-input/index.js
@@ -452,6 +452,7 @@ class URLInput extends Component {
 			value,
 			required: true,
 			type: 'text',
+			name: inputId,
 			onChange: disabled ? () => {} : this.onChange, // Disable onChange when disabled
 			onFocus: disabled ? () => {} : this.onFocus, // Disable onFocus when disabled
 			placeholder,