+++ packages/block-library/src/form/index.js
@@ -1,4 +1,9 @@
 /**
+ * WordPress dependencies
+ */
+import { addFilter } from '@wordpress/hooks';
+
+/**
  * Internal dependencies
  */
 import initBlock from '../utils/init-block';
@@ -7,17 +12,14 @@ import metadata from './block.json';
 import save from './save';
 import variations from './variations';
 import deprecated from './deprecated';
-
-/**
- * WordPress dependencies
- */
-import { addFilter } from '@wordpress/hooks';
+import { icon } from './icons';
 
 const { name } = metadata;
 
 export { metadata, name };
 
 export const settings = {
+	icon,
 	edit,
 	save,
 	deprecated,