+++ packages/sync/src/providers/index.ts
@@ -6,32 +6,18 @@ import { applyFilters } from '@wordpress/hooks';
 /**
  * Internal dependencies
  */
-import { createIndexedDbProvider } from './indexeddb-provider';
-import { createWebRTCProvider } from './webrtc-provider';
 import type { ProviderCreator } from '../types';
 
 let providerCreators: ProviderCreator[] | null = null;
 
 /**
- * Returns provider creators for IndexedDB and WebRTC with HTTP signaling. These
- * are the current default providers.
+ * Returns Yjs provider creators for collaborative editing. There are currently
+ * no default providers, so this function returns an empty array.
  *
  * @return {ProviderCreator[]} Creator functions for Yjs providers.
  */
 function getDefaultProviderCreators(): ProviderCreator[] {
-	const signalingUrl = window?.wp?.ajax?.settings?.url;
-
-	if ( ! signalingUrl ) {
-		return [];
-	}
-
-	return [
-		createIndexedDbProvider,
-		createWebRTCProvider( {
-			password: window?.__experimentalCollaborativeEditingSecret,
-			signaling: [ signalingUrl ],
-		} ),
-	];
+	return [];
 }
 
 /**