+++ packages/theme/bin/terrazzo-plugin-known-wpds-css-variables/index.ts
@@ -5,7 +5,7 @@ import { FORMAT_ID } from '@terrazzo/plugin-css';
 import type { Plugin } from '@terrazzo/parser';
 
 export default function pluginKnownWpdsCssVariables( {
-	filename = 'design-tokens.js',
+	filename = 'design-tokens.mjs',
 } = {} ): Plugin {
 	return {
 		name: '@wordpress/terrazzo-plugin-known-wpds-css-variables',