+++ packages/theme/src/prebuilt/ts/color-tokens.ts
@@ -10,7 +10,6 @@ export default {
 		'fg-interactive-brand-strong-active',
 	],
 	'primary-bgFill2': [ 'bg-interactive-brand-strong-active' ],
-	'primary-surface2': [ 'bg-interactive-brand-active' ],
 	'primary-surface4': [ 'bg-interactive-brand-weak-active' ],
 	'primary-fgSurface3': [
 		'fg-interactive-brand',
@@ -72,23 +71,8 @@ export default {
 	'error-stroke4': [ 'stroke-interactive-error-active' ],
 	'error-stroke1': [ 'stroke-surface-error' ],
 	'bg-surface2': [ 'bg-surface-neutral' ],
-	'bg-surface6': [
-		'bg-interactive-brand-strong-disabled',
-		'bg-interactive-error-strong-disabled',
-		'bg-interactive-neutral-strong-disabled',
-	],
-	'bg-surface5': [
-		'bg-interactive-brand-disabled',
-		'bg-interactive-brand-weak-disabled',
-		'bg-interactive-error-disabled',
-		'bg-interactive-error-weak-disabled',
-		'bg-interactive-neutral-disabled',
-		'bg-interactive-neutral-weak-disabled',
-	],
-	'bg-surface4': [
-		'bg-interactive-neutral-active',
-		'bg-interactive-neutral-weak-active',
-	],
+	'bg-surface5': [ 'bg-interactive-neutral-strong-disabled' ],
+	'bg-surface4': [ 'bg-interactive-neutral-weak-active' ],
 	'bg-surface3': [ 'bg-surface-neutral-strong' ],
 	'bg-fgSurface4': [
 		'fg-content-neutral',
@@ -97,15 +81,11 @@ export default {
 	],
 	'bg-fgSurface3': [
 		'fg-content-neutral-weak',
-		'fg-interactive-brand-strong-disabled',
-		'fg-interactive-error-strong-disabled',
-		'fg-interactive-neutral-strong-disabled',
 		'fg-interactive-neutral-weak',
 	],
 	'bg-fgSurface2': [
-		'fg-interactive-brand-disabled',
-		'fg-interactive-error-disabled',
 		'fg-interactive-neutral-disabled',
+		'fg-interactive-neutral-strong-disabled',
 		'fg-interactive-neutral-weak-disabled',
 	],
 	'bg-stroke3': [
@@ -119,7 +99,7 @@ export default {
 		'stroke-interactive-neutral-strong',
 	],
 	'bg-stroke2': [
-		'bg-thumb-brand-disabled',
+		'bg-thumb-neutral-disabled',
 		'bg-track-neutral',
 		'stroke-interactive-neutral-disabled',
 		'stroke-surface-neutral',