+++ storybook/preview.jsx
@@ -11,10 +11,8 @@ import {
  * Internal dependencies
  */
 import { WithGlobalCSS } from './decorators/with-global-css';
-import { WithMarginChecker } from './decorators/with-margin-checker';
 import { WithMaxWidthWrapper } from './decorators/with-max-width-wrapper';
 import { WithRTL } from './decorators/with-rtl';
-import { WithTheme } from './decorators/with-theme';
 import { WithDesignSystemTheme } from './decorators/with-design-system-theme';
 
 export const globalTypes = {
@@ -27,20 +25,6 @@ export const globalTypes = {
 			items: [
 				{ value: 'ltr', title: 'LTR' },
 				{ value: 'rtl', title: 'RTL' },
-			],
-		},
-	},
-	componentsTheme: {
-		name: 'Theme',
-		description: 'Change the components theme. (Work in progress)',
-		defaultValue: 'default',
-		toolbar: {
-			icon: 'paintbrush',
-			items: [
-				{ value: 'default', title: 'Default' },
-				{ value: 'darkBg', title: 'Dark (background)' },
-				{ value: 'lightGrayBg', title: 'Light gray (background)' },
-				{ value: 'classic', title: 'Classic (accent)' },
 			],
 		},
 	},
@@ -58,19 +42,6 @@ export const globalTypes = {
 					value: 'wordpress',
 					title: 'WordPress (common, forms, dashicons)',
 				},
-			],
-		},
-	},
-	marginChecker: {
-		name: 'Margin Checker',
-		description:
-			'Show a div before and after the component to check for unwanted margins.',
-		defaultValue: 'hide',
-		toolbar: {
-			icon: 'collapse',
-			items: [
-				{ value: 'hide', title: 'Hide' },
-				{ value: 'show', title: 'Show' },
 			],
 		},
 	},
@@ -94,16 +65,17 @@ export const globalTypes = {
 
 export const decorators = [
 	WithGlobalCSS,
-	WithMarginChecker,
 	WithRTL,
 	WithMaxWidthWrapper,
-	WithTheme,
 	WithDesignSystemTheme,
 ];
 
 export const parameters = {
 	controls: {
 		sort: 'requiredFirst',
+	},
+	backgrounds: {
+		disable: true,
 	},
 	docs: {
 		controls: {
@@ -159,6 +131,8 @@ export const parameters = {
 					'Typography',
 				],
 				'Icons',
+				'Design System',
+				[ 'Introduction', 'Theme', 'Components', [ 'Introduction' ] ],
 			],
 		},
 	},