+++ packages/keycodes/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New features
+
+-   Add new `ariaKeyShortcut` function ([#74205](https://github.com/WordPress/gutenberg/pull/74205)).
+
 ## 4.36.0 (2025-11-26)
 
 ## 4.35.0 (2025-11-12)